// frontend/generated/models/ProposalStatusDto.ts
/* tslint:disable */
/* eslint-disable */
/**
 * Freelance Platform API
 * No description provided (generated by Openapi Generator https://github.com/openapitools/openapi-generator)
 *
 * The version of the OpenAPI document: 1.0.0
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */


/**
 * 
 * @export
 */
export const ProposalStatusDto = {
    Pending: 'PENDING',
    Expired: 'EXPIRED',
    Withdrawn: 'WITHDRAWN',
    Accepted: 'ACCEPTED',
    Rejected: 'REJECTED'
} as const;
export type ProposalStatusDto = typeof ProposalStatusDto[keyof typeof ProposalStatusDto];


export function instanceOfProposalStatusDto(value: any): boolean {
    for (const key in ProposalStatusDto) {
        if (Object.prototype.hasOwnProperty.call(ProposalStatusDto, key)) {
            if (ProposalStatusDto[key as keyof typeof ProposalStatusDto] === value) {
                return true;
            }
        }
    }
    return false;
}

export function ProposalStatusDtoFromJSON(json: any): ProposalStatusDto {
    return ProposalStatusDtoFromJSONTyped(json, false);
}

export function ProposalStatusDtoFromJSONTyped(json: any, ignoreDiscriminator: boolean): ProposalStatusDto {
    return json as ProposalStatusDto;
}

export function ProposalStatusDtoToJSON(value?: ProposalStatusDto | null): any {
    return value as any;
}

export function ProposalStatusDtoToJSONTyped(value: any, ignoreDiscriminator: boolean): ProposalStatusDto {
    return value as ProposalStatusDto;
}